refactor and change folder structure
