+++ api/__init__.py
@@ -1,0 +1,6 @@
+# API init 
+
+# API package
+from api.routes import create_router
+
+__all__ = ["create_router"]
