+++ graphs/__init__.py
@@ -1,0 +1,8 @@
+# Graphs init 
+
+# Graphs module
+
+from graphs.compliance import create_compliance_graph
+from graphs.feasibility import build_feasibility_graph
+
+__all__ = ["create_compliance_graph", "build_feasibility_graph"]
